requirements.txt
# PxGAN Requirements
# Core dependencies for Pixelation-Guided GAN

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0              # Parquet support
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Logging and experiment tracking
tensorboard>=2.13.0
# Optional: wandb, mlflow (uncomment if needed)
# wandb>=0.15.0
# mlflow>=2.5.0

# Utilities
scipy>=1.10.0               # For statistical functions
joblib>=1.3.0               # For caching and parallel processing

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Optional GPU acceleration (uncomment if using CUDA)
# nvidia-ml-py3>=7.352.0     # For GPU monitoring

# Code quality (development)
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
